requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
idna==2.10
importlib-metadata==1.7.0
more-itertools==8.4.0
packaging==20.4
pathlib2==2.3.5
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pymongo==3.11.0
pyparsing==2.4.7
pytest==5.4.3
requests==2.24.0
selenium==3.141.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.5
zipp==1.2.0








